+++ .gitignore
@@ -1,1 +1,1 @@
-# data file*data/* 
+# data file*data/* *posters/*.Rproj.user
