--- visuals/et_viz.py
@@ -1,1 +1,0 @@
-!pip install plotlyfrom os import chdir, getcwdimport pandas as pdimport matplotlib.pyplot as pltimport seaborn as snsimport plotly.express as pximport plotly.io as piopio.renderers.default='browser'wd=getcwd() # set working directorychdir(wd)# read in dfet_means = pd.read_csv(wd + '/data/analysis_results/et_df_1381151.csv', index_col=0)et_means = et_means[['time', 'disalexi',                      'ensemble', 'geesebal',                      'ptjpl', 'sims', 'ssebop', 'Et']]et_means['year'] = pd.to_datetime(et_means['time']).dt.strftime('%Y')# plot line graph for the OpenET and AquaCrop  et_means.plot(x = 'time', linewidth=1, alpha=0.4)plt.plot(et_means['time'], et_means['Et'], color='black')plt.xticks(rotation=90)  plt.ylabel('ET (mm)')plt.xlabel('')plt.legend(bbox_to_anchor=(1.02, 1), loc='upper left', borderaxespad=0)# et model comparison plotet_val = et_means.drop(columns = ['time']) # drop time variable# plot looks funky, so removed 0 values from Etet_val.plot(x = 'Et', linewidth=1, alpha=0.4)#plt.plot(et_means['time'], et_means['Et'], color='black')plt.xticks(rotation=90)  plt.ylabel('OpenET (mm)')plt.xlabel('AquaCrop ET (mm)')plt.legend(bbox_to_anchor=(1.02, 1), loc='upper left', borderaxespad=0)# plot without the zero values from the aquacrop modelet_val_nonzero = et_val[et_val['Et'] > 0]et_val_nonzero.plot(x = 'Et', linewidth=1, alpha=0.4)#plt.plot(et_means['time'], et_means['Et'], color='black')plt.xticks(rotation=90)  plt.ylabel('OpenET (mm)')plt.xlabel('AquaCrop ET (mm)')plt.legend(bbox_to_anchor=(1.02, 1), loc='upper left', borderaxespad=0)#x = sns.FacetGrid(et_val2, col='year', col_wrap=3, )#x = x.map(plt.plot, 'disalexi',                      #'ensemble', 'geesebal',                      #'ptjpl', 'sims', 'ssebop', 'Et')#plt.show()# create new df and remove AquaCrop Et values < 0 then facet_warp by yearet_val2 = et_val[et_val['Et'] > 0]et_val2 = pd.melt(et_val2, id_vars=['year','Et'],                  var_name='model_name', value_name='values')fig = px.line(et_val2, x="Et", y="values", color='model_name',                 facet_col="year",                 labels={                     "Et": "Aquacrop Et (mm)",                     "values": "OpenET (mm)",                     "model_name": "Model Name"                 })fig.show() ## not looking good...
