+++ visuals_code/irrigation_viz.py
@@ -1,1 +1,1 @@
-!pip install sklearnfrom os import chdir, getcwdimport pandas as pdimport matplotlib.pyplot as pltfrom sklearn.metrics import mean_squared_error from matplotlib.ticker import MaxNLocatorfrom matplotlib import patchesimport seaborn as snswd=getcwd() # set working directorychdir(wd)# read in dfyield_df = pd.read_csv(wd + '/data/analysis_results/yield_df_1381151.csv', index_col=0)irrig_df = pd.read_csv(wd + '/data/analysis_results/irrig_df_1381151.csv', index_col =0)#boxplot showing wimas irrigation and irrigation when irrig is trigered different soil moisture threshlos#yield 2006-2015# irrigation colors#print(sns.color_palette("Blues_d").as_hex())irrig_colors = ['#FFA500', '#8abbdb', '#69a6d0', '#4892c6', '#367bac', '#356384', '#344b5b']# get the upper and lower quatile values for the WIMAS irrigation . Will use to add horizontal stripirrig_df.irrig_wimas.quantile([0.25,0.5,0.75])sns.set(font_scale = 1.5)sns.set_style("white")fig, ax = plt.subplots(figsize=(10, 8))ax = sns.boxplot(data=irrig_df,                  color = 'black',                 palette=sns.color_palette(irrig_colors),                 linewidth=1.2,                  fliersize=2,                  order=['irrig_wimas', '0', '20', '40', '60', '80', '100'],                 flierprops=dict(marker='o', markersize=4)) #ax.tick_params(bottom=False, left=True)#ax.ayvspan(min_irrig, max_irrig, color='grey', alpha=0.5, lw=0)rect = patches.Rectangle(    xy=(ax.get_xlim()[0], 330),  # lower left corner of box: beginning of x-axis range & y coord)    width=ax.get_xlim()[1]-ax.get_xlim()[0],  # width from x-axis range    height=170,    color='grey', alpha=0.2,    zorder=1) #ec='red'ax.add_patch(rect)ax.set_xticklabels(['WIMAS', '0', '20', '40', '60', '80', '100'])#ax.set_yticklabels(ax.get_yticks(), y.astype(int), size = 15)ax.set_xlabel('Soil Moisture Threshold (%TAW)', size=25, weight='bold')ax.set_ylabel('Irrigation (mm)', size = 25, weight='bold')fig.savefig('results/visuals/wimas_aquacrop_boxplot_irrigation_soilMoisture.png', format='png', dpi=500)#boxplot with USDS NASS yield and Aquacrop yield 2000-2014# get the upper and lower quatile values for the AquaCrop yield at the 80% SMT. Will use to add horizontal stripyield_df['80'].describe()#print(sns.light_palette("seagreen").as_hex())yield_colors = ['#FFA500', '#ebf3ed', '#c5decf', '#9fc9b1', '#7ab493', '#54a075', '#2e8b57']sns.set(font_scale = 1.5)sns.set_style("white")fig, ax = plt.subplots(figsize=(10, 8))bplot = sns.boxplot(data=yield_df,                  color = 'black',                 palette=sns.color_palette(yield_colors),                 linewidth=1.2,                  fliersize=2,                  order=['YieldUSDA', '0', '20', '40', '60', '80', '100'],                 flierprops=dict(marker='o', markersize=4)) #ax.tick_params(bottom=False, left=True)rect = patches.Rectangle(    xy=(ax.get_xlim()[0], 13.8),  # lower left corner of box: beginning of x-axis range & y coord)    width=ax.get_xlim()[1]-ax.get_xlim()[0],  # width from x-axis range    height=.7,    color='grey', alpha=0.2) #ec='red'ax.add_patch(rect)         ax.set_xticklabels(['USDA', '0', '20', '40', '60', '80', '100'])ax.set_xlabel('Soil Moisture Threshold (%TAW)', size=25, weight='bold')ax.set_ylabel('Yield (t/ha)', size = 25, weight='bold')fig.savefig('results/visuals/wimas_aquacrop_boxplot_yield_soilMoisture.png', format='png', dpi=500)
+!pip install sklearnfrom os import chdir, getcwdimport pandas as pdimport matplotlib.pyplot as pltfrom sklearn.metrics import mean_squared_error from matplotlib.ticker import MaxNLocatorfrom matplotlib import patchesimport seaborn as snswd=getcwd() # set working directorychdir(wd)# read in dfyield_df = pd.read_csv(wd + '/data/analysis_results/yield_df_1381151.csv', index_col=0)irrig_df = pd.read_csv(wd + '/data/analysis_results/irrig_df_1381151.csv', index_col =0)#boxplot showing wimas irrigation and irrigation when irrig is trigered different soil moisture threshlos#yield 2006-2015# irrigation colors#print(sns.color_palette("Blues_d").as_hex())irrig_colors = ['#FFA500', '#8abbdb', '#69a6d0', '#4892c6', '#367bac', '#356384', '#344b5b']# get the upper and lower quatile values for the WIMAS irrigation . Will use to add horizontal stripirrig_df.irrig_wimas.quantile([0.25,0.5,0.75])sns.set(font_scale = 1.5)sns.set_style("white")fig, ax = plt.subplots(figsize=(10, 8))ax = sns.boxplot(data=irrig_df,                  color = 'black',                 palette=sns.color_palette(irrig_colors),                 linewidth=1.2,                  fliersize=2,                  order=['irrig_wimas', '0', '20', '40', '60', '80', '100'],                 flierprops=dict(marker='o', markersize=4)) #ax.tick_params(bottom=False, left=True)#ax.ayvspan(min_irrig, max_irrig, color='grey', alpha=0.5, lw=0)rect = patches.Rectangle(    xy=(ax.get_xlim()[0], 330),  # lower left corner of box: beginning of x-axis range & y coord)    width=ax.get_xlim()[1]-ax.get_xlim()[0],  # width from x-axis range    height=170,    color='grey', alpha=0.2,    zorder=1) #ec='red'ax.add_patch(rect)ax.set_xticklabels(['WIMAS', '0', '20', '40', '60', '80', '100'])#ax.set_yticklabels(ax.get_yticks(), y.astype(int), size = 15)ax.set_xlabel('Soil Moisture Threshold (%TAW)', size=25, weight='bold')ax.set_ylabel('Irrigation (mm)', size = 25, weight='bold')fig.savefig('results/visuals/wimas_aquacrop_boxplot_irrigation_soilMoisture_aquacropSiltLoam.png', format='png', dpi=500)#boxplot with USDS NASS yield and Aquacrop yield 2000-2014# get the upper and lower quatile values for the AquaCrop yield at the 80% SMT. Will use to add horizontal stripyield_df['80'].describe()#print(sns.light_palette("seagreen").as_hex())yield_colors = ['#FFA500', '#ebf3ed', '#c5decf', '#9fc9b1', '#7ab493', '#54a075', '#2e8b57']sns.set(font_scale = 1.5)sns.set_style("white")fig, ax = plt.subplots(figsize=(10, 8))bplot = sns.boxplot(data=yield_df,                  color = 'black',                 palette=sns.color_palette(yield_colors),                 linewidth=1.2,                  fliersize=2,                  order=['YieldUSDA', '0', '20', '40', '60', '80', '100'],                 flierprops=dict(marker='o', markersize=4)) #ax.tick_params(bottom=False, left=True)rect = patches.Rectangle(    xy=(ax.get_xlim()[0], 13.8),  # lower left corner of box: beginning of x-axis range & y coord)    width=ax.get_xlim()[1]-ax.get_xlim()[0],  # width from x-axis range    height=.7,    color='grey', alpha=0.2) #ec='red'ax.add_patch(rect)         ax.set_xticklabels(['USDA', '0', '20', '40', '60', '80', '100'])ax.set_xlabel('Soil Moisture Threshold (%TAW)', size=25, weight='bold')ax.set_ylabel('Yield (t/ha)', size = 25, weight='bold')fig.savefig('results/visuals/wimas_aquacrop_boxplot_yield_soilMoisture.png', format='png', dpi=500)
